Added chart
